Fixed error where Maprows delete button not found on options page

+++ public/script.js
@@ -146,6 +146,7 @@ document.addEventListener("DOMContentLoaded", function () {
     });
   }
 
+  //MAP Row sort buttons
   // Mileage and Pay Method
   let currentMapRowsMethod = "all";
 
@@ -184,30 +185,33 @@ document.addEventListener("DOMContentLoaded", function () {
   });
 
   // MAP Rows delete button
-  document.getElementById("MapRowsData").addEventListener("click", async function (event) {
-    if (event.target.matches(".delete-btn")) {
-      const button = event.target;
-      const id = button.getAttribute("data-id");
-      const date = button.getAttribute("data-date");
-      const totalMiles = button.getAttribute("data-total-miles");
-      const pay = button.getAttribute("data-pay");
+  if (document.getElementById("MapRowsData")) {
+    document.getElementById("MapRowsData").addEventListener("click", async function (event) {
+      if (event.target.matches(".delete-btn")) {
+        const button = event.target;
+        const id = button.getAttribute("data-id");
+        const date = button.getAttribute("data-date");
+        const totalMiles = button.getAttribute("data-total-miles");
+        const pay = button.getAttribute("data-pay");
 
-      try {
-        // Send delete request
-        const response = await fetch(`/deleteRow/${id}`, {
-          method: "DELETE",
-          headers: {
-            "Content-Type": "application/json",
-          },
-          body: JSON.stringify({ date, totalMiles, pay }),
-        });
-        fetchMapRowsData(highlightMapRowsButton);
-      } catch (error) {
-        console.error("Error:", error);
-        alert("An error occurred while deleting the row.");
+        try {
+          const response = await fetch(`/deleteRow/${id}`, {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+            },
+            body: JSON.stringify({ date, totalMiles, pay }),
+          });
+          fetchMapRowsData(highlightMapRowsButton);
+          fetchSortedPartial(currentIntervalSortMethod);
+        } catch (error) {
+          console.error("Error:", error);
+          alert("An error occurred while deleting the row.");
+        }
       }
-    }
-  });
+    });
+  }
+
   // insertCompanyForm
   if (document.getElementById("companyForm")) {
     document.getElementById("companyForm").addEventListener("submit", function (event) {
